refactor(database): split module boundaries
