+++ src/middleware/database/raw.hpp
@@ -4,567 +4,15 @@
 #include <cstring>
 
 #include "flash.hpp"
-#include "libxr.hpp"
-#include "libxr_def.hpp"
-#include "libxr_mem.hpp"
-#include "libxr_type.hpp"
+#include "interface.hpp"
 
 namespace LibXR
 {
 
-static constexpr uint16_t LIBXR_DATABASE_VERSION = 3;
-
-/**
- * @brief 数据库接口，提供键值存储和管理功能 (Database interface providing key-value
- * storage and management).
- */
-class Database
-{
- public:
-  /**
-   * @brief 键的基类，存储键名及其数据 (Base class for keys, storing key name and
-   * associated data).
-   */
-  class KeyBase
-  {
-   public:
-    const char* name_;  ///< 键名 (Key name).
-    RawData raw_data_;  ///< 原始数据 (Raw data associated with the key).
-
-    /**
-     * @brief 构造函数，初始化键名和原始数据 (Constructor to initialize key name and raw
-     * data).
-     * @param name 键名 (Key name).
-     * @param raw_data 关联的原始数据 (Raw data associated with the key).
-     */
-    KeyBase(const char* name, RawData raw_data) : name_(name), raw_data_(raw_data) {}
-  };
-
-  /**
-   * @brief 模板类，表示数据库中的具体键 (Template class representing a specific key in
-   * the database).
-   * @tparam Data 存储的数据类型 (The type of data stored).
-   */
-  template <typename Data>
-  class Key : public KeyBase
-  {
-   public:
-    Data data_;           ///< 键存储的数据 (The data stored in the key).
-    Database& database_;  ///< 关联的数据库对象 (Reference to the associated database).
-
-    /**
-     * @brief 构造函数，初始化键并从数据库加载数据 (Constructor to initialize key and load
-     * data from the database).
-     *
-     * If the key does not exist in the database, it is initialized with the provided
-     * value. 如果键在数据库中不存在，则使用提供的值进行初始化。
-     *
-     * @param database 关联的数据库对象 (Reference to the associated database).
-     * @param name 键名 (Key name).
-     * @param init_value 初始化值 (Initial value for the key).
-     */
-    Key(Database& database, const char* name, Data init_value)
-        : KeyBase(name, RawData(data_)), database_(database)
-    {
-      ErrorCode status = database.Get(*this);
-      if (status != ErrorCode::OK)
-      {
-        data_ = init_value;
-        if (status == ErrorCode::NOT_FOUND)
-        {
-          REQUIRE(database.Add(*this) == ErrorCode::OK);
-        }
-      }
-    }
-
-    /**
-     * @brief 构造函数，初始化键，并在数据库不存在时赋默认值 (Constructor to initialize
-     * key, assigning default value if not found in the database).
-     *
-     * If the key does not exist in the database, it is initialized with zero.
-     * 如果键在数据库中不存在，则初始化为零。
-     *
-     * @param database 关联的数据库对象 (Reference to the associated database).
-     * @param name 键名 (Key name).
-     */
-    Key(Database& database, const char* name)
-        : KeyBase(name, RawData(data_)), database_(database)
-    {
-      ErrorCode status = database.Get(*this);
-      if (status != ErrorCode::OK)
-      {
-        Memory::FastSet(&data_, 0, sizeof(Data));
-        if (status == ErrorCode::NOT_FOUND)
-        {
-          REQUIRE(database.Add(*this) == ErrorCode::OK);
-        }
-      }
-    }
-
-    /**
-     * @brief 禁止拷贝数据库键对象 (Copy construction is disabled for database keys).
-     * @param other 被拷贝的键对象 (Database key to copy from).
-     */
-    Key(const Key&) = delete;
-
-    /**
-     * @brief 禁止移动数据库键对象 (Move construction is disabled for database keys).
-     * @param other 被转移的键对象 (Database key to move from).
-     */
-    Key(Key&&) = delete;
-
-    /**
-     * @brief 禁止拷贝赋值数据库键对象
-     *        (Copy assignment is disabled for database keys).
-     * @param other 被拷贝的键对象 (Database key to copy from).
-     * @return 当前键对象引用 (Reference to the current key object).
-     */
-    Key& operator=(const Key&) = delete;
-
-    /**
-     * @brief 禁止移动赋值数据库键对象
-     *        (Move assignment is disabled for database keys).
-     * @param other 被转移的键对象 (Database key to move from).
-     * @return 当前键对象引用 (Reference to the current key object).
-     */
-    Key& operator=(Key&&) = delete;
-
-    /**
-     * @brief 类型转换运算符，返回存储的数据 (Type conversion operator returning stored
-     * data).
-     * @return 存储的数据 (Stored data).
-     */
-    operator Data() { return data_; }
-
-    /**
-     * @brief 保存当前键值到数据库 (Save the current key value to the database).
-     * @return 操作结果 (Operation result).
-     */
-    ErrorCode Save() { return database_.Set(*this, this->raw_data_); }
-
-    /**
-     * @brief 设置键的值并更新数据库 (Set the key's value and update the database).
-     * @param data 需要存储的新值 (New value to store).
-     * @return 操作结果 (Operation result).
-     */
-    ErrorCode Set(Data data)
-    {
-      data_ = data;
-      return Save();
-    }
-
-    /**
-     * @brief 从数据库加载键的值 (Load the key's value from the database).
-     * @return 操作结果 (Operation result).
-     */
-    ErrorCode Load() { return database_.Get(*this); }
-
-    /**
-     * @brief 赋值运算符，设置键的值 (Assignment operator to set the key's value).
-     * @param data 需要存储的新值 (New value to store).
-     * @return 操作结果 (Operation result).
-     */
-    ErrorCode operator=(Data data) { return Set(data); }  // NOLINT
-  };
-
- private:
-  /**
-   * @brief 从数据库获取键的值 (Retrieve the key's value from the database).
-   * @param key 需要获取的键 (Key to retrieve).
-   * @return 操作结果 (Operation result).
-   */
-  virtual ErrorCode Get(KeyBase& key) = 0;
-
-  /**
-   * @brief 设置数据库中的键值 (Set the key's value in the database).
-   * @param key 目标键 (Target key).
-   * @param data 需要存储的新值 (New value to store).
-   * @return 操作结果 (Operation result).
-   */
-  virtual ErrorCode Set(KeyBase& key, RawData data) = 0;
-
-  /**
-   * @brief 添加新键到数据库 (Add a new key to the database).
-   * @param key 需要添加的键 (Key to add).
-   * @return 操作结果 (Operation result).
-   */
-  virtual ErrorCode Add(KeyBase& key) = 0;
-};
-
-/**
- * @brief 适用于不支持逆序写入的 Flash 存储的数据库实现
- *        (Database implementation for Flash storage that does not support reverse
- * writing).
- *
- * This class manages key-value storage in a Flash memory region where
- * data can only be written sequentially. It maintains a backup system
- * to prevent data corruption.
- * 此类管理 Flash 内存区域中的键值存储，其中数据只能顺序写入。
- * 它维护一个备份系统，以防止数据损坏。
- *
- * @note 若底层 Flash 读写擦失败，当前实现视为不可恢复故障并直接触发 `REQUIRE`。
- *       If the underlying Flash read, write, or erase operation fails, the
- *       current implementation treats it as an unrecoverable fault and triggers
- *       `REQUIRE` immediately.
- */
-class DatabaseRawSequential : public Database
-{
- public:
-  /**
-   * @brief 构造函数，初始化 Flash 存储和缓冲区
-   *        (Constructor initializing Flash storage and buffer).
-   *
-   * @param flash 目标 Flash 存储设备 (Target Flash storage device).
-   * @param max_buffer_size 最大缓冲区大小，默认 256 字节 (Maximum buffer size, default is
-   * 256 bytes).
-   *
-   * @note 包含动态内存分配。
-   *       Contains dynamic memory allocation.
-   * @note `max_buffer_size` 必须不超过整片 Flash 容量的一半。
-   *       `max_buffer_size` must not exceed half of the total flash capacity.
-   */
-  explicit DatabaseRawSequential(Flash& flash, size_t max_buffer_size = 256);
-
-  /**
-   * @brief 初始化数据库存储区，确保主备块正确
-   *        (Initialize database storage, ensuring main and backup blocks are valid).
-   */
-  void Init();
-
-  /**
-   * @brief 保存当前缓冲区内容到 Flash
-   *        (Save the current buffer content to Flash).
-   */
-  void Save();
-
-  /**
-   * @brief 从 Flash 加载数据到缓冲区
-   *        (Load data from Flash into the buffer).
-   */
-  void Load();
-
-  /**
-   * @brief 还原存储数据，清空 Flash 区域
-   *        (Restore storage data, clearing Flash memory area).
-   */
-  void Restore();
-
-  /**
-   * @brief 获取数据库中的键值
-   *        (Retrieve the key's value from the database).
-   * @param key 需要获取的键 (Key to retrieve).
-   * @return 操作结果，如果找到则返回 `ErrorCode::OK`，否则返回 `ErrorCode::NOT_FOUND`
-   *         (Operation result, returns `ErrorCode::OK` if found, otherwise
-   * `ErrorCode::NOT_FOUND`).
-   */
-  ErrorCode Get(Database::KeyBase& key) override;
-
-  /**
-   * @brief 设置数据库中的键值
-   *        (Set the key's value in the database).
-   * @param key 目标键 (Target key).
-   * @param data 需要存储的新值 (New value to store).
-   * @return 操作结果 (Operation result).
-   */
-  ErrorCode Set(KeyBase& key, RawData data) override
-  {
-    return SetKey(key.name_, data.addr_, data.size_);
-  }
-
-  /**
-   * @brief 添加新键到数据库
-   *        (Add a new key to the database).
-   * @param key 需要添加的键 (Key to add).
-   * @return 操作结果 (Operation result).
-   */
-  ErrorCode Add(KeyBase& key) override
-  {
-    return AddKey(key.name_, key.raw_data_.addr_, key.raw_data_.size_);
-  }
-
- private:
-  /**
-   * @brief 存储块类型 (Storage block type).
-   */
-  /**
-   * @brief 存储块类型 (Storage block type).
-   */
-  /**
-   * @brief 存储块类型 (Storage block type).
-   */
-  enum class BlockType : uint8_t
-  {
-    MAIN = 0,   ///< 主块 (Main block).
-    BACKUP = 1  ///< 备份块 (Backup block).
-  };
-
-#pragma pack(push, 1)
-  /**
-   * @brief 键信息结构，存储键的元数据
-   *        (Structure containing key metadata).
-   */
-  struct KeyInfo
-  {
-    uint32_t raw_data;
-
-    /**
-     * @brief 构造一个擦除态键头 (Construct one erased-state key header).
-     */
-    KeyInfo() : raw_data(0xFFFFFFFF) {}
-
-    /**
-     * @brief 构造一个指定元数据的键头 (Construct one key header with explicit metadata).
-     * @param nextKey 是否还有后继键 (Whether another key follows this one).
-     * @param nameLength 键名长度 (Key name length).
-     * @param dataSize 数据字节数 (Payload size in bytes).
-     */
-    KeyInfo(bool nextKey, uint8_t nameLength, uint32_t dataSize) : raw_data(0)
-    {
-      SetNextKeyExist(nextKey);
-      SetNameLength(nameLength);
-      SetDataSize(dataSize);
-    }
-
-    /**
-     * @brief 设置是否存在后继键 (Set whether another key follows).
-     * @param value 是否存在后继键 (Whether another key follows).
-     */
-    void SetNextKeyExist(bool value)
-    {
-      raw_data = (raw_data & 0x7FFFFFFF) | (static_cast<uint32_t>(value & 0x1) << 31);
-    }
-
-    /**
-     * @brief 获取是否存在后继键 (Get whether another key follows).
-     * @return 若存在后继键则返回 `true` (Returns `true` when another key follows).
-     */
-    bool GetNextKeyExist() const { return (raw_data >> 31) & 0x1; }
-
-    /**
-     * @brief 设置键名长度 (Set the key name length).
-     * @param len 键名长度 (Key name length).
-     */
-    void SetNameLength(uint8_t len)
-    {
-      raw_data = (raw_data & 0x80FFFFFF) | (static_cast<uint32_t>(len & 0x7F) << 24);
-    }
-
-    /**
-     * @brief 获取键名长度 (Get the key name length).
-     * @return 键名长度 (Key name length).
-     */
-    uint8_t GetNameLength() const { return (raw_data >> 24) & 0x7F; }
-
-    /**
-     * @brief 设置数据字节数 (Set the payload size in bytes).
-     * @param size 数据字节数 (Payload size in bytes).
-     */
-    void SetDataSize(uint32_t size)
-    {
-      raw_data = (raw_data & 0xFF000000) | (size & 0x00FFFFFF);
-    }
-
-    /**
-     * @brief 获取数据字节数 (Get the payload size in bytes).
-     * @return 数据字节数 (Payload size in bytes).
-     */
-    uint32_t GetDataSize() const { return raw_data & 0x00FFFFFF; }
-  };
-
-  static_assert(sizeof(KeyInfo) == 4, "KeyInfo size must be 4 bytes");
-#pragma pack(pop)
-
-  /**
-   * @brief Flash 存储的块信息结构
-   *        (Structure representing a Flash storage block).
-   */
-  struct FlashInfo
-  {
-    uint32_t header;  ///< Flash 块头标识 (Flash block header identifier).
-    KeyInfo key;      ///< 该块的键信息 (Key metadata in this block).
-  };
-
-  /**
-   * @brief 按名称新增一个键 (Add one key by name).
-   * @param name 键名 (Key name).
-   * @param data 键数据地址 (Address of the key payload).
-   * @param size 键数据字节数 (Payload size in bytes).
-   * @return 操作结果 (Operation result).
-   */
-  ErrorCode AddKey(const char* name, const void* data, size_t size);
-
-  /**
-   * @brief 按名称更新一个键 (Update one key by name).
-   * @param name 键名 (Key name).
-   * @param data 键数据地址 (Address of the key payload).
-   * @param size 键数据字节数 (Payload size in bytes).
-   * @return 操作结果 (Operation result).
-   */
-  ErrorCode SetKey(const char* name, const void* data, size_t size);
-
-  /**
-   * @brief 按存储偏移更新一个键 (Update one key by storage offset).
-   * @param offset 键头偏移 (Key-header offset).
-   * @param data 键数据地址 (Address of the key payload).
-   * @param size 键数据字节数 (Payload size in bytes).
-   * @return 操作结果 (Operation result).
-   */
-  ErrorCode SetKey(size_t offset, const void* data, size_t size);
-
-  /**
-   * @brief 读取一个键的数据区 (Read one key payload).
-   * @param offset 键头偏移 (Key-header offset).
-   * @param data 接收数据的缓冲区 (Destination buffer receiving the payload).
-   * @return 操作结果 (Operation result).
-   */
-  ErrorCode GetKeyData(size_t offset, RawData data);
-
-  /**
-   * @brief 把指定块初始化为空数据库块 (Initialize one block as an empty database block).
-   * @param block 目标块类型 (Target block type).
-   */
-  void InitBlock(BlockType block);
-
-  /**
-   * @brief 读取 Flash 数据，失败则直接触发强约束
-   *        (Read flash data and fail fast on error).
-   * @param offset 读取偏移 (Read offset).
-   * @param data 接收数据的缓冲区 (Destination buffer receiving the data).
-   */
-  void ReadFlashOrExit(size_t offset, RawData data);
-
-  /**
-   * @brief 读取 Flash 数据到对象里，失败则直接触发强约束
-   *        (Read flash data into one object and fail fast on error).
-   * @tparam Data 接收对象类型 (Destination object type).
-   * @param offset 读取偏移 (Read offset).
-   * @param data 接收数据的对象 (Destination object receiving the data).
-   */
-  template <typename Data>
-  void ReadFlashOrExit(size_t offset, Data& data)
-  {
-    ReadFlashOrExit(offset, RawData(data));
-  }
-
-  /**
-   * @brief 写入 Flash 数据，失败则直接触发强约束
-   *        (Write flash data and fail fast on error).
-   * @param offset 写入偏移 (Write offset).
-   * @param data 待写入数据 (Data to write).
-   */
-  void WriteFlashOrExit(size_t offset, ConstRawData data);
-
-  /**
-   * @brief 擦除 Flash 区域，失败则直接触发强约束
-   *        (Erase a flash range and fail fast on error).
-   * @param offset 擦除偏移 (Erase offset).
-   * @param size 擦除字节数 (Erase size in bytes).
-   */
-  void EraseFlashOrExit(size_t offset, size_t size);
-
-  /**
-   * @brief 判断块头是否已初始化 (Check whether the block header is initialized).
-   * @param block 目标块类型 (Target block type).
-   * @return 若块头有效则返回 `true` (Returns `true` when the block header is valid).
-   */
-  bool IsBlockInited(BlockType block);
-
-  /**
-   * @brief 判断块当前是否为空 (Check whether the block is currently empty).
-   * @param block 目标块类型 (Target block type).
-   * @return 若块内没有有效键则返回 `true`
-   *         (Returns `true` when the block contains no valid key).
-   */
-  bool IsBlockEmpty(BlockType block);
-
-  /**
-   * @brief 判断块尾校验是否损坏 (Check whether the block checksum is corrupted).
-   * @param block 目标块类型 (Target block type).
-   * @return 若块尾校验不符则返回 `true`
-   *         (Returns `true` when the trailing checksum is invalid).
-   */
-  bool IsBlockError(BlockType block);
-
-  /**
-   * @brief 判断指定键后面是否还有下一键 (Check whether another key follows).
-   * @param offset 当前键头偏移 (Current key-header offset).
-   * @return 若后面还有下一键则返回 `true`
-   *         (Returns `true` when another key follows this one).
-   */
-  bool HasLastKey(size_t offset);
-
-  /**
-   * @brief 计算一个键总共占用的字节数 (Compute the total byte span of one key).
-   * @param offset 键头偏移 (Key-header offset).
-   * @return 该键占用的总字节数 (Total byte size occupied by the key).
-   */
-  size_t GetKeySize(size_t offset);
-
-  /**
-   * @brief 计算下一键的起始偏移 (Compute the starting offset of the next key).
-   * @param offset 当前键头偏移 (Current key-header offset).
-   * @return 下一键的起始偏移 (Starting offset of the next key).
-   */
-  size_t GetNextKey(size_t offset);
-
-  /**
-   * @brief 计算当前块里最后一个键的偏移 (Locate the last key in the current block).
-   * @param block 目标块类型 (Target block type).
-   * @return 最后一个键的偏移；若块为空则返回 `0`
-   *         (Offset of the last key, or `0` when the block is empty).
-   */
-  size_t GetLastKey(BlockType block);
-
-  /**
-   * @brief 回写上一键的“存在后继键”标志 (Rewrite the next-key-exists flag of one key).
-   * @param offset 目标键头偏移 (Target key-header offset).
-   * @param exist 是否存在后继键 (Whether another key follows).
-   */
-  void SetNestKeyExist(size_t offset, bool exist);
-
-  /**
-   * @brief 比较存储中的键数据和给定数据是否不同
-   *        (Compare whether the stored payload differs from the given payload).
-   * @param offset 键头偏移 (Key-header offset).
-   * @param data 待比较数据地址 (Address of the candidate payload).
-   * @param size 待比较数据字节数 (Payload size in bytes).
-   * @return 若内容不同则返回 `true`
-   *         (Returns `true` when the payloads differ).
-   */
-  bool KeyDataCompare(size_t offset, const void* data, size_t size);
-
-  /**
-   * @brief 比较存储中的键名和给定名称是否不同
-   *        (Compare whether the stored key name differs from the given name).
-   * @param offset 键头偏移 (Key-header offset).
-   * @param name 待比较键名 (Key name to compare against).
-   * @return 若名称不同则返回 `true`
-   *         (Returns `true` when the names differ).
-   */
-  bool KeyNameCompare(size_t offset, const char* name);
-
-  /**
-   * @brief 在主块里按名称查找键 (Search one key by name in the main block).
-   * @param name 待查找键名 (Key name to search for).
-   * @return 找到时返回键头偏移，找不到返回 `0`
-   *         (Returns the key-header offset when found, otherwise `0`).
-   */
-  size_t SearchKey(const char* name);
-
-  static constexpr uint32_t FLASH_HEADER =
-      0x12345678 + LIBXR_DATABASE_VERSION;  ///< Flash 头部标识 (Flash header identifier).
-  static constexpr uint8_t CHECKSUM_BYTE = 0x56;  ///< 校验字节 (Checksum byte).
-
-  Flash& flash_;              ///< 目标 Flash 存储设备 (Target Flash storage device).
-  uint8_t* buffer_;           ///< 数据缓冲区 (Data buffer).
-  uint32_t block_size_;       ///< Flash 块大小 (Flash block size).
-  uint32_t max_buffer_size_;  ///< 最大缓冲区大小 (Maximum buffer size).
-};
-
 /**
  * @brief 适用于最小写入单元受限的 Flash 存储的数据库实现
  *        (Database implementation for Flash storage with minimum write unit
- * restrictions).
+ *        restrictions).
  *
  * This class provides key-value storage management for Flash memory that
  * requires data to be written in fixed-size blocks.
@@ -576,7 +24,7 @@ class DatabaseRawSequential : public Database
  *       `REQUIRE` immediately.
  *
  * @tparam MinWriteSize Flash 的最小写入单元大小 (Minimum write unit size for Flash
- * storage).
+ *         storage).
  */
 template <size_t MinWriteSize>
 class DatabaseRaw : public Database
